+++ RELEASE-NOTES.md
@@ -1,4 +1,5 @@
 * 1.7.6
+   * Add J8Streams.unfoldResource inspired by [akka-streams Source.unfoldResource](http://doc.akka.io/docs/akka/2.4/scala/stream/stages-overview.html#unfoldResource).
    * Switch to optional findbugs:jsr305 to not pollute others with the transitive dep on annotations and for `@ParametersAreNonnullByDefault`.
 * 1.7.5
    * Added J8Futures for converting between [CompletableFuture](https://docs.oracle.com/javase/8/docs/api/java/util/concurrent/CompletableFuture.html) and Guava's [ListenableFuture](http://google.github.io/guava/releases/snapshot/api/docs/com/google/common/util/concurrent/ListenableFuture.html)